第七次提交

+++ src/main/java/com/fc/service/impl/DormitoryManagerServiceImpl.java
@@ -1,10 +1,14 @@
 package com.fc.service.impl;
 
+import com.fc.dao.BuildingMapper;
 import com.fc.dao.DormitoryManagerMapper;
+import com.fc.entity.Building;
+import com.fc.entity.BuildingExample;
 import com.fc.entity.DormitoryManager;
 import com.fc.entity.DormitoryManagerExample;
 import com.fc.service.DormitoryManagerService;
 import com.fc.util.RandomNum;
+import com.fc.vo.DataVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +19,9 @@ import java.util.List;
 public class DormitoryManagerServiceImpl implements DormitoryManagerService {
     @Autowired
     private DormitoryManagerMapper dormitoryManagerMapper;
+
+    @Autowired
+    private BuildingMapper  buildingMapper;
 
     @Override
     public DormitoryManager findByName(String name) {
@@ -83,7 +90,25 @@ public class DormitoryManagerServiceImpl implements DormitoryManagerService {
     }
 
     @Override
-    public void delete(String id) {
-        dormitoryManagerMapper.deleteByPrimaryKey(id);
+    public DataVO delete(String id) {
+        BuildingExample buildingExample = new BuildingExample();
+
+        BuildingExample.Criteria criteria = buildingExample.createCriteria();
+
+        criteria.andDormitoryManagerIdEqualTo(id);
+
+        List<Building> buildings = buildingMapper.selectByExample(buildingExample);
+
+        DataVO dataVO = null;
+
+        if (buildings.size() == 0) {
+            dormitoryManagerMapper.deleteByPrimaryKey(id);
+
+            dataVO = new DataVO(200, "开除成功", true);
+        } else {
+            dataVO = new DataVO(400, "该宿管已有任务，请先找人接替，在开除", false);
+        }
+
+        return dataVO;
     }
 }
